fix: Fix stylelint reporting issues coming from prettier
This is not a real fix of the problem itself, its more like a sticking plaster
since at the moment I dont have a better idea. Switching the execution order of
Prettier and Stylelint in the lint-fix script helps at least with the ugly line-
breaks coming from Prettiers Print Width rule, that cannot be disabled.

#2